Handling allocation domain of the input TensorViews in the matmul scheduler  (#2309)
In this PR we extend the matmul scheduler to support inputs with
allocation domains.


To the fusion (with inputs tv_a and tv_b), we add two LoadStoreOps to
both inputs.
The first Op corresponds to a load to shared memory, where we propagate
the allocation domain. The second op corresponds to reading to
registers, where we don't propagate the allocation domain since the
scheduler takes charge of setting the allocation domain in the
registers. Based on the difference in the (maybe)allocation domain of
the producer and consumer of the second LoadStoreOp, we may do
transposed load when reading to registers.


![image](https://github.com/NVIDIA/Fuser/assets/10635897/89395990-9b85-4ce1-8e7d-006e43a86b85)


See also https://github.com/NVIDIA/Fuser/pull/2315.