Fix coverage with multiple expr groups producing the same val group (#2322)
Previously, in `computeCoveredGroups`, when `covered_ids` has already a
mapping, the additional covered groups are simply ignored. This could
cause the loop promotion analysis to fail to find any promotion. See the
added test for a concrete example. That repro is rather synthetic, but
the same issue also happens with some of the reshape shmoo tests.