[User Schedule] Implement transform_like and parallelize_like (#2285)
This PR implements the `transform_like` and `parallelize_like`
functions.

`transform_like(a: Tensor, selected_nodes : [Tensor] = [] )` propagates
the iterDomain transformations to the rest of the fusion. It is a
wrapper around `TransformPropagator` and
`MaxRootDomainInfoSpanningTree`.

When the `selected_nodes` argument is not empty, it will use the
`SetSelector` to apply the transformations to a subset of the graph.

```cpp
TransformPropagator propagator(reference_tv);
SetSelector selector({subgraph_tv_set.begin(), subgraph_tv_set.end()});
MaxRootDomainInfoSpanningTree(reference_tv, &selector).traverse(&propagator);
```

`parallelize_like` is a direct wrapper around
`nvfuser::scheduler_utils::parallelizeAllLike`, which will propagate the
parallelization scheme to the rest of the fusion.