Convert from dynamic type to static type early (#2325)
When talking about design decisions in `DynamicType`, there are some
subtle details that is hard to tell which is the best decision. For
example, `at::Tensor == at::Tensor` returns an `at::Tensor`, but some
C++ code assumes `operator==` always returns `bool`. For this case,
should we make `PolymorphicValue == PolymorphicValue` always return
`bool` and disable `at::Tensor == at::Tensor`, or make it return
`PolymorphicValue` which is a superset of `bool` and can be casted to
`bool`? There is no clear answer to this question. None is perfect, it's
just a matter of choosing which bad you want to take. So far, I believe
the decision is based on whatever the behavior is that makes
`DynamicType` codebase simple.

Recently, I added `DynamicType::dispatch`, which is primarily for
supporting https://github.com/NVIDIA/Fuser/pull/2321, but I also noticed
that this is a powerful tool that can be used to cleanup implementations
of `DynamicType`'s operators.
To be able to more thoroughly cleaning things up, some decisions like
above could change. Again, I believe this change is neither an
improvement nor a regression on the behavior. It is just a different
behavior.

One pitfall for using `PolymorphicValue` is, its `operator&&` and
`operator||` are no longer short-circuiting. This is a fundamental
limitation of C++ operator overloading, and to the best of my knowledge,
there is no workaround for it. This means, if you write code like below:
```C++
pv.hasValue() && f(pv)
```
what you are thinking is likely
```C++
if (pv.hasValue()) {
  return (bool)f(pv);
} else {
  return false;
}
```
but what you get will be
```C++
PolymorphicValue cond1 = pv.hasValue();
PolymorphicValue cond2 = f(pv);
return (bool)(cond1 && cond2);
```
That is, `f(pv)` is always evaluated and likely this evaluation will
leads to errors. To get the correct behavior, you will have to use
```C++
pv.hasValue() && (bool)(f(pv))
```
so that the `operator&&` is the `operator&&` of `bool`, not of
`PolymorphicValue`.

In the past, I changed a batch of patterns like this during the
enablement of `PolymorphicValue`, but there are still patterns like
above remain unchanged because it was saved by the fact that the
comparison operator of `PolymorphicValue` returning `bool`. If I changed
the comparison operators to return `PolymorphicValue`, these uses will
break too.

This PR further changes codes like above, so that we always use
`operator&&` of bool regardless of the return type of the comparison
operators.