Don't invalidate existing ValGroup when not necessary (#2345)
Currently, functions in `id_model/schedule.{h,cpp}` will invalidate the
input `ValGroup` because it will create a new `ValGroup` with a new ID
and map that new group with the input group. This is annoying and
unnecessarily expensive. This PR changes the behavior by just appending
the new ID into the existing `ValGroup`.