Initial PR of IdModel-based indexing (#2344)
This is the first of a series of IdModel-based indexing PRs. The overall
algorithm is exactly what I presented at the nvfuser meeting and is
based on the shortest path traversal introduced at #2270. This PR only
includes minimum logic for basic indexing. However, pointwise and
reduction ops with TID/BID parallelization as well as reshape should
work.

Notable missing support:

- Broadcast
- Vectorize, Unswitch, Unroll
- Contiguous indexing

Note that the new indexing interface is not plugged into the lowering
yet. Currently, it's only used by the new unit tests. In subsequent PRs,
there'll be a switch to use the new indexer instead of the existing one.

Overall tracking PR: #2238