Add small repro as python test
This test is actually tougher than the big repro on #2359. It
necessitate either removing the check altogether or the path I took
which is to concretize Resized to broadcast extents as constant 1 so
that we can evaluate max(0, min(i0, 1)) as 1 without calling
simplifyExpr. A less invasive solution would be to remove the extent
check in `SqueezeOp::checkConcretization`.

We could also just remove `Expr::checkConcretization` and
`checkConcretizedUses`. They are only used for SqueezeOp currently and
are not adding much value anyway probably.