Host irs: use `Communicator` directly instead of wrapped inside `PostOnStream` (#2328)
Bootstrapping for simplifying the design. With this patch, the
`Communication` class is used to post a collective, without the need of
wrapping it inside a `PostOnStream` Ir.

There is no functional change, this is only a minor cleanup I would like
to upstream before introducing the "collective wait" IR.

To achieve this, we need to add I/O TVs to the `Communication`'s
constructor, and use those I/O directly. This patch aligns with the
discussion from https://github.com/NVIDIA/Fuser/pull/2250 and should
lead to further simplification in `MultiDeviceExecutor`, that I would
like to leave to another PR for avoiding conflicts.

Note that the `Communication` class is used in two places: Host IRs and
`MultiDeviceExecutor`. Since the former case should replace the latter
eventually, we maintain for now these two usages. In
`MultiDeviceExecutor`, we don't set a `Communication`'s I/O, that's why
those arguments are made optional in the constructor

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>