Support indexing of vectorized tensors (#2380)
Just supporting indexing with vectorization. Note that this is still
only for indexing used in tensor ops themselves. Predicate indexing will
be supported later.

There's only one thing that needs to be done, which is to use zero for a
vectorized domain. This could be done by `shouldUseZeroIndex`, but for
now I decided to generate an index and then replace the corresponding
val with zero. This is because vectorized domains do not always just use
zero but `N-1`, where `N` is the extent of the domain, for predication.