Add uniform/normal for half/bfloat (#2355)
This PR implements uniform and normal sampling for half and bfloat
types. This is done by sampling a 32bit uint, casting to float, dividing
by 2^32 and adding 2^{-33}, then casting to half/bfloat. After the cast,
we check whether the result is equal to 1 and if so, we return 0. This
matches pytorch's approach to the best of my knowledge (see
https://github.com/pytorch/pytorch/blob/c59a2369bec99dbbc088ac5477efe1ab6adc7ec8/aten/src/ATen/native/cuda/DistributionTemplates.h#L475).

Fixes #1653