Change getProblemLayout to getOperandInnerDims (#2422)
This is related to #2419 which is working toward generalizing our
scheduler to handle any number of operands instead of exactly 2.

In order to keep the change small, I introduced a `MmaLayout
getMatmulProblemLayout(Fusion*)` utility in `tests/cpp/utils.cpp`. This
should give the same result that `mma_utils::getProblemLayout(fusion)`
used to give, but asserting instead of wrapping in a `DataWrapperOpt`,
so we don't need to change our supported layouts much. For new code, we
could change to support vectors of inner dimension roles instead.

After this PR, `MmaLayout` should only be used in these places
- In `MmaOpUtils::getInputLayout` which finds the layout of the actual
inputs to the `MmaOp` (not the Fusion inputs). This is the correct
location to use `MmaLayout` since `MmaOp` supports a single A operand
and a single B operand.
- In C++ matmul tests, which still use `MmaLayout` to describe the
problem layout.
- In C++ matmul benchmarks, which similarly use `MmaLayout`.