Add prefix option to the communicator's backend cache (#2402)
# What
Adds a fine grain control of the Communicator's backend cache

# Why
The `Communicator` is used as a singleton instance managing the
different backends for communication. It internally holds a cache of the
previously created backend to avoid the costly initial overhead of
ProcessGroup creation.

One the one hand, it is important that this caching occurs at the
Communicator's level and not at the caller site, because the
communicator exists globally accross test instances.

On the other hand, in order to achieve communication/communication
overlap, one must use multiple backends because each backend has one and
only one internal stream on which it posts kernels.

# How
We add an optional `std::string prefix=""` argument to the function
`Communicator:: getBackendForTeam` that is prepended to the key used to
retrieve a backend from the cache. This allows arbitrary control of the
cache from the caller.