Refactor lowering of TMA (#2413)
This PR refactors the lowering of TMA into two stages:
1. An analysis that computes the ValGroups for all the roles (tile
ValGroup, box ValGroup, partitioned ValGroup, etc.)
2. An index lowering that uses the result of 1 to get the index.

I intentionally keep the algorithm unchanged. The analysis is almost a
copy-paste of the original `getCpAsyncBulkGmemIndex`, except that it is
modified to traverse on the exact graph, instead of on the gmem_tv
replayed as consumer. In the future, I might do more cleanup to better
take advantage of IdModel.

In step 2, we are using the legacy indexing, instead of the IdModel
based indexing. I intentionally not change this so that this PR is
effectively a no-op. In the future, the legacy indexing will be replaced
with IdModel based indexing.

cc: @rdspring1 @protonu