host ir: add support for for-loop (#2440)
# What 
This PR introduces the support for for-loop used as host IRs.

# Test
We provide a unit test that reproduces in host IR the program
```
int64_t buf = kInitialValue;
for (int j = kForLoopStart; j < kForLoopStop; j += kForLoopStep) {
   buf += j;
}
```
where buf is the global output.

# remark on kir and hir namespace

We reuse the `kir::ForLoop` to avoid code duplication and to reuse in
the future the utils already implemented there. I know there is a
conflict in the namespace, and we shouldn't use "kir" in "hir". I don't
know what's the best practice here: either "kir::ForLoop" and
"hir::ForLoop" should inherit from the same base class, or we both IRs
should diverge.

I see no fundamental reason to diverge for now, though, expect that the
constructor expect an `IterDomain` which is not used in our case for now
(but that could change in the future).

@wujingyue wdyt?

# remark on using FusionExecutor

Our test cannot run with `FusionExecutor` but only with
`FusionExecutorCache`. The reason is kind of incidental and is related
to concretization. Indeed, when `HostIrExecutor` uses `FusionExecutor`,
it does not perform any concretization on the `Fusion`. This problem is
known and present in `MultiDeviceExecutor` -- since it was not a
priority to add concretization, the workaround was only to ensure
nothing is dynamic in the Fusion. In the present case, I have not
figured out how to perform this with the "iota" IR, and am getting this
error if I attempt to use FusionExecutor:
> dynamic_tvs.empty() INTERNAL ASSERT FAILED at
/opt/pytorch/Fuser/csrc/device_lower/validation.cpp:187, please report a
bug with repro script to NVFuser at
https://github.com/NVIDIA/Fuser/issues. Tensor with dynamic transform
must be concretized before lowering: T1_l[ ?S2{( ( ( -( fmax(0, (
where(( i7 < 0 ), ( i7 + 10 ), i7) )) ) ) + 10 ) + ( ( fmax(( fmax(0, (
where(( i7 < 0 ), ( i7 + 10 ), i7) )) ), ( fmin(10, ( where(( ( i7 + 1 )
< 0 ), ( ( i7 + 1 ) + 10 ), ( i7 + 1 )) )) )) ) - 10 ) )}rf ], T3_g[
?S4{( ( ( -( fmax(0, ( where(( i7 < 0 ), ( i7 + 10 ), i7) )) ) ) + 10 )
+ ( ( fmax(( fmax(0, ( where(( i7 < 0 ), ( i7 + 10 ), i7) )) ), (
fmin(10, ( where(( ( i7 + 1 ) < 0 ), ( ( i7 + 1 ) + 10 ), ( i7 + 1 )) ))
)) ) - 10 ) )} ]

I believe this error is not important and is completely orthogonal to
the PR, that's why in the present patch we skip the failing test.

This being said, I also believe adding fusion concretization shouldn't
be too hard (especially if we don't cache re-compilation), but I haven't
figured out how to do that yet.