Canonicalize matmul dims in scheduleMatmul using dimension roles (#2376)
This updates `scheduleMatmul` to use the dimension roles introduced in
#2303 to reorder and merge dims within each role. That means we can
schedule fusions with more than one tensor in each role.

Two tests are included so far:
- MultipleMDims which performs [ M1, M2, K] @ [N, K]. This corresponds
to `torch.linear` with 3D input i.e. a Linear layer with two "batch"
dimensions.
- MultipleMDimsBatch which performs [M1, B, M2, K] @ [B, N, K]. This
shows that M dimensions need not be contiguous. Note that vectorization
is unaffected in this example since K is innermost.
I plan to add more tests to explore more scenarios.

Note that we cannot yet handle cases where K is not the innermost
dimension, and I have not yet tested cases with M as the innermost
output dimension. These will likely require us to modify the fusion
definition to place those dimensions in the right spot in the MmaOp.

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>