Allow domains to be decoupled: validateDomainEquivalence (#2379)
In the past, it was assumed that domains in a `TensorDomain` has linear
structure `root -> logical -> allocation -> leaf`. I am removing this
restriction. As the step 0 of the removal of this restriction, I am
modifying the `validateDomainEquivalence`, which is called in the
constructor of `TensorDomain`, to support arbitrary dependency order.