Simplify handling of "define box by compositing" (#2420)
When there is a "define box by compositing", in principle, we should be
indexing into the _imagined_ partitioned IterDomain, as shown below"


![](https://raw.githubusercontent.com/NVIDIA/Fuser/main/doc/dev/tma/box-by-compositing.svg)

However, we were not able to do that in the past because indexing only
supports backward propagation. With `TensorIndexer` and
`AbstractTensor`, we can easily create the _imagined_ partitioned
ValGroup and index into it.
So there is no need to use the state machine approach to manually
combine indices, instead, the index of the _imagined_ partitioned
ValGroup can be directly used as the index for the TMA PTX instructions.