Comms/compute Overlap with RS-Based pipelining: ATen implementation (#2401)
# What

This PR provides an ATen implementation of "GEMM+Reduce-scatter"
operation using a reduce-scattered based pipelining overlapping
technique, as used in NEMO-Megatron transformer, precisely at the second
layer of the MLP.

# Configurations

We provide a way to fully configure the parameters used in the test
through environment variables. The available options are:
- The dimensions `M,K,N` of the matrices
- The number of interleaved chunks `S`
- Whether to use several streams
- which communication backend to use, NCCL or UCC

# Detailed walkthrough

The tensor program that we target is the following, assuming a setup
with `num_devices` devices:
   - inputs:
- `A[M,K]` sharded column-wise: dimension `K` is split by the factor
`num_devices`, so `A` is viewed as `[M, num_devices, K/num_devices]`,
and the allocation size of `A` is `[M, 1, K/num_devices]`
- `B[K,N]` sharded row-wise: locally of size `[1, K/num_devices, N]`
   - output:
- `C[M,N]=matmul(A,B)`, sharded on dimension `M`: dimension `M` is split
by the factor `num_devices`, so `C` is viewed as `[num_devices,
M/num_devices,N]`, and the allocation size of `M` is `[1,
M/num_devices,N]`

Up to some broadcast and view ops, a straightforward program to generate
the output could be summarized as
```
C_unreduced = pointwise_multiply(A,B)
C_locally_reduce = reduction(C_unreduced, axis=`K/num_devices`, op=sum)
C = reduce_scatter(C_unreduced, op=sum)
```
where `C_unreduced`, `C_locally_reduce`, `C` are all sharded on
`num_devices` and :
- `C_unreduced` has unsharded size `[M,num_devices,K/num_devices,N]`,
- `C_locally_reduce` has unsharded size `[M,num_devices,N]`,
- `C` has unsharded size `[num_devices, M/num_devices, N]`


We want to compare this baseline program with one that is functionally
identical but achieves more overlap between computations and
communications. Our goal is to interlave the comms and compute using a
technic called "reduce-scatter based pipelining". To do so, we further
split the row dimension `M` with a factor `S` and we apply the comms and
compute described earlier in parallel (i.e. on different streams) on
each tensors slices accross `S`. More precisely, assuming the following
shapes:
    - `A [S, num_devices, M/S, K/num_devices]`, sharded on `num_devices`
    - `B [num_devices, K/num_devices, N]`, sharded on `num_devices`
- `C [S, num_devices, M/(S*num_devices), N]`, sharded on `num_devices`
the program achieving comms/compute overlap using rs-based pipelining
could be summarized as:
```
for (j=0; j<S; j++) {
    setCurrentStream(Stream[j])
    C_unreduced[j] = pointwise_multiply(A[j],B)
    C_locally_reduce[j] = local_reduction(C_unreduced[j], axis=`K/num_devices`, op=sum)
    C[j]=reduce_scatter(C_locally_reduce[j], op=sum)
}
```
where `[j]` refers to taking a slice onto the `S` dimension.

_Remarks:_
1) it is convenient to have `S` as being the outermost dimension so
`C_locally_reduce[j]` is a contiguous buffer, and can therefore be
reduce-scattered with a single MPI collective.
2) The layout needs to match the reduce-scatter semantics, i.e., the
first dimension is reduced and the second is scattered. This is why we
choose the layouts to be`[S, sharded_axis, M / (...) , ...]`


# NSight profile analysis

### Setup 
We provide a screenshot of the night profile, obtained with the
following setup:
- 8 GPUs V100, on one DGX node
- M=N=K=2^10
- S=8
- Multi stream activated
- backend: nccl

### Profile analysis
Comments:
- We can see on the illustration the three types of kernels:
   - nccl Kernels for reduce_scatter
   - pointwise kernels (for the pointwise multiply)
   - reduce kernels (for the local sum reduction)
- (Circle1)The first kernel we see on the screenshot is the first
reduce-scatter. The first compute, happening before reduce-scatter,
cannot be seen on the screenshot because of the nccl init overhead.
Also, the nccl init is blocking, which explains why the other compute
are not hidden by nccl init.
- (Circle2) I _think_ that ATen compute takes all the SMs and therefore
nccl Kernel is enqueued, hence the non-overlap.
- (Circle3) During reduce-scatter, ATen compute overlaps. We see
comm/compute and compute/compute overlap.
- (Circle4) For the same reason as in Circle 2, nccl kernels are
enqueued to after the compute is finished
- (Circle4) The comms do not overlap one with another because PG has
only one internal stream. Upcoming optimization: instantiating multiple
PG backends to achieve comm/comm overlap. Dependency with
https://github.com/NVIDIA/Fuser/pull/2402 and an upcoming PR
implementing a dedicated Host IR
![Screenshot 2024-06-13 at 17 07
17](https://github.com/NVIDIA/Fuser/assets/17732757/ab2db696-c872-4226-b39e-9a92a662e9ad)