Refactor `swizzleSharedMemory` to return an `AbstractTensor` (#2467)
After this PR, `swizzleSharedMemory` will no longer modify the tensor
that should be swizzled. Instead, it returns an `AbstractTensor` that
contains the swizzled domain. For now, this domain must be set as the
loop domain of the swizzled tensor.

In the future, I will modify `swizzleSharedMemory` to use the new
`Swizzle` instead of the deprecated `Swizzle2D`. The returned domain of
`swizzleSharedMemory` will be the allocation domain instead of the loop
domain of the swizzled tensor.
When everything is done, `Swizzle2D` will be removed.