Adding unit tests to check TMA loads for input B (#2468)
In the new unit test added we load the input B tensor to shared memory
using TMA. These tests are for a single tile, where M=64, N=8 ... 256
and K=16. So for the TN layout B: {N, K} TT: {K, N}. In the following
figure we show how we tile the tensor to 4 loads.


![image](https://github.com/NVIDIA/Fuser/assets/10635897/e1c1bb7e-6644-40a6-bed7-0b5d659920bb)

The figure below shows the scheduling for the above tiling. This
schedule/tiling works for all inputs sizes (and TT and TN)


![image](https://github.com/NVIDIA/Fuser/assets/10635897/38e4583e-48d7-4056-83f3-fa0b6318f13a)

In addition to the above tiling/shedules, we could have scheduled the
same shape above to reduce the number of TMA loads to 2 from 4. It's
illustrated in the figure below. Please note that it correctly works for
cases where we don't need to split the inner dimension. As future work
we can try and make it work for all cases.


![image](https://github.com/NVIDIA/Fuser/assets/10635897/738a9aa4-7324-48d0-b2b2-659b1e7a3b63)

The schedule for the above tiling scheme is show below. This is
illustrated in the test `SingleTileWithTMALoadOuterDimNotSplit`

![image](https://github.com/NVIDIA/Fuser/assets/10635897/c2f777e5-75f2-4402-a495-45f88c5f409b)