Fix for when allocation domains are promoted to reduction domains (#2500)
For non-global tensors, unless explicitly set, allocation domains are
promotion domains of loop domains. Promotion can be a reduction domain
even for non-reduction domains as there's no special handling for
reduction domains. This can result in incorrect indexing with the
current code. Specifically, we skip reduction domains from allocation,
however, since we look at promotion domains, they may be reduction
domains even though original loop domains are not. This can result in
incorrectly skipping indexing. See the added unit test for a concrete
example.

The fix in this PR is to use both the original loop domain and its
promotion domain when determining the allocation domain of a given
tensor to index.