HostIrContainer: add default stream (#2491)
# What
add in `HostIrContainer` an IR representation of the "default stream" as
defined by torch

# Why

used in the overlap implementation

https://github.com/NVIDIA/Fuser/blob/3258fa5d3bf9e89c600c1615f6dbc24b00b8dcce/tests/cpp/test_multidevice_overlap.cpp#L220

# How

`IrContainer::oneVal()` is a similar concept as what we try to achieve
here. But the implementation there has some complexity

https://github.com/NVIDIA/Fuser/blob/3258fa5d3bf9e89c600c1615f6dbc24b00b8dcce/csrc/ir/container.cpp#L266
in particular, `oneVal` is not built with `IrBuilder::create` as one
would expect, though I am not sure to understand the benefit.

Here, I didn't care to reproduce the implementation of `oneVal`, I
propose something simpler relying on `IrBuilder::create`. Let me know if
you prefer something in the same vein as the implementation of `oneVal`.