patch isZero/isOne (#2510)
1. There's no overload across equal operator for `std::complex<T>, T`
```
auto x = std::complex<float>(0.0, 0.0);
x == 0; // This one triggers an error.
```

So we hit this with our dynamic type thing in `Val::isZero()`. But it
magically worked if I change it from an integer to a float. I have no
idea why. cc'ing @zasdfgbnm

2. `at::pad` does not support complex dtype. But nvfuser API does. Added
a chain of operation to support it.