Support indexing of unrolled domains (#2516)
Unrolling affects allocations, so indexing also needs to be aware of
unrolled domains. More specifically, the allocation position is the
minimum of the CA position and unroll domain.

Furthermore, for certain tensors, just looking at its loop domains may
not be sufficient if the tensor is broadcast and inlined. For example,

```
 auto tv0 = makeContigConcreteTensor({4});
  fusion.addInput(tv0);
  auto tv1 = makeContigConcreteTensor({3, 4});
  fusion.addInput(tv1);

  auto tv2 = set(tv0);
  auto tv3 = broadcast(tv2, {true, false});
  auto tv4 = add(tv3, tv1);
  fusion.addOutput(tv4);

  tv4->split(0, 1);

  TransformPropagator propagator(tv4);
  MaxRootDomainInfoSpanningTree(tv4).traverse(&propagator);

  inlineMost();

  tv4->axis(1)->parallelize(ParallelType::Unroll);
```

```
Inputs:
  T0_g[ iS0{4} ], float
  T1_g[ iS10{( ceilDiv(3, 1) )}, iS11{1}, iS2{4} ], float
Outputs:
  T4_g[ iS8{( ceilDiv(3, 1) )}, iUR9{1}, iS7{4} ] ca_pos( 3 ) produce_pos( 3 ), float

%kernel_math {
T2_l[ iS3{4} ] ca_pos( 1 )
   = Set( T0_g[ iS0{4} ], cache_op=Streaming )
T3_l[ bS12{( ceilDiv(1, 1) )}, bUR13{1}, iS5{4} ] ca_pos( 3 ) produce_pos( 3 )
   = broadcast( T2_l[ iS3{4} ] ca_pos( 1 ) )
T4_g[ iS8{( ceilDiv(3, 1) )}, iUR9{1}, iS7{4} ] ca_pos( 3 ) produce_pos( 3 )
   = T3_l[ bS12{( ceilDiv(1, 1) )}, bUR13{1}, iS5{4} ] ca_pos( 3 ) produce_pos( 3 )
   + T1_g[ iS10{( ceilDiv(3, 1) )}, iS11{1}, iS2{4} ];
}
```

See `InlinedUnroll` for the full detail.

In this case, `tv2` is fully inlined into `tv3` and then `tv4`. Notice
that while `tv4` is unrolled, but just looking at `tv2` doesn't allow
you to find the unrolling. In other words, the loop domains of `tv2`
fail to represent the unrolling.

I think this is yet another instance where we may want to do something
like [this](#2459). If we inserted a broadcast domain to `tv2` when it's
inlined, that broadcast domain would capture the unrolling.

While I do think broadcast insertion is a better approach, for now, I'm
fixing it by checking the for-loops where a tensor is defined. The logic
of finding the correct allocation position considering inlined unrolled
domains is already
[there](https://github.com/NVIDIA/Fuser/blob/main/csrc/device_lower/utils.h#L285),
so I'm just using it for now.

Additionally, I moved `getAllocationDomains` to a separate class and
changed the overall flow such that before indexing, traversal of KIR
exprs must be done to build a map from each tensor to its allocation
information. Previously, the allocation information is collected on the
fly, but it didn't work well now that for-loops at the time of
definitions are required.