Reorder allocation domains to align with logical domains (#2522)
I put why this needs to be done
[here](https://github.com/NVIDIA/Fuser/pull/2522/files#diff-0c47d4d562dba268498573ced9dd7fc7d9638b5882aca4c4cc20e70a7ccec891R315).
For a concrete example, see the added unit test.

Alternatively, we could make loop vectorized domains always placed at
the rightmost position. Since the indexing scheme uses the loop domains
to infer the allocation domain, that would solve the issue as well. For
now, I'm adding this patch so that the new indexer can work with the
existing schedulers with no change.

Stacked on top of #2516