Patch for allocation domains of shmem tensors used for transpose (#2536)
(stacked on #2523)

The new indexing framework starts with identifying (or inferring)
allocation domains of each tensor. The logic is similar to what's done
in the allocation lowering pass, and indeed some of the logic is reused
in indexing.

One case where the inference fails to identify is the required order of
allocation domains. One instance is addressed in #2522. Another instance
is when a shared memory is used as a staging buffer for storing a
transposed tensor. Please see the code comment of
`patchAllocationOfTransposedSmemTensor`.

While this should be good enough for now, I'm wondering if we should
actually set the allocation domain of the staging tensor explicitly.
When scheduling the fusion, we know what the allocation domain should
look like. What's done here is to infer the domain from the fusion
pattern. This approach doesn't require any change in the scheduler but
the inference may not be always correct.

Let me know what you think, @zasdfgbnm