Rewrite TMA analysis (#2557)
In the current main branch, the analysis in
`csrc/device_lower/analysis/tma.cpp` mostly comes from copy-pasting the
legacy indexing code of TMA with simple string replacement to fit for
IdGraph. In this approach, although we are using IdGraph, I still don't
consider it modern, because we are still assuming the traversal path to
be like a TensorDomain (For example, traversal is only backward, no
forward traversal is allowed, and each val group only have one
definition).

This PR rewrites the TMA analysis to make it modern. This rewrite makes
the analysis more extendable so that in the future I can easily add
support for "define box by rotation". As a side effect of this rewrite,
we now only requires a path from consumer's loop domain to gmem tv's
allocation domain, and it does not matter if the traversal is forward or
backward. This PR does not add any new unit tests, but in the next PR, I
will add a few unit tests for testing backward+forward traversal.

`tma.cpp` is almost completely rewritten, so I do not recommend looking
at the diff. I suggest totally discard the original code and read the
new code.

cc: @rdspring1 @protonu