Sharded SDPAFwdOp (#2565)
Temporarily add support for a sharded forward scaled dot product
attention.
Currently, we only support DID parallelization on the logical domain
which requires us to split and parallelize an axis at the logical level
see https://github.com/NVIDIA/Fuser/issues/2563. This is a hack until we
support DID parallelization on the loop domain after which this PR can
be reverted.

Restrictions:
1. q,k,v inputs are manually sharded _before_ the SDPAFwdOp is created.
We cannot rely on sharding propagation or sharding after the Fusion is
created, because the dimension checks are called when the op is created.
2. Only the head dimension is sharded and all inputs and outputs have
either a sharded head dimension or unshaded.
3. DID axis is the outermost axis. This is because during evaluation if
we see 5 dimensions, it is assumed the first is the DID axis and is
appropriately squeezed from the inputs and unsqueezed onto the outputs.